### 2024-04-08 #6592 @editorialbot
[PRE REVIEW]: IonDiff: command-line tool to identify ionic diffusion events and hopping correlations in molecular dynamics simulations
**Submitting author:** @IonRepo (<a href="http://orcid.org/0000-0002-6501-4513">Claudio Cazorla</a>)
**Repository:** https://github.com/IonRepo/IonDiff
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
