### 2024-05-02 #6616 @LIVazquezS
Dear authors,

Here are some comments about your work.

## General Checks

### Contribution and authorship

Could the authors clarify the contributions of authors Riccardo Rurali and Claudio Cazorla? Are they sharing the git account IonRepo?

### Functionality

- The authors must automatize the installation of the software. I could not install it by just following the procedure described.
- The authors claimed that their implementation scales linearly with different factors, such as the length of the temporal window, the total duration of the simulation, and the number of mobile ions. Evidence for this claim must be provided.
- The authors mention vectorizing operations using a common library such as Numpy. However, could this be improved by using Numba? Is the code parallelized?
- The authors claim that their implementation outperforms standard evaluation algorithms; evidence for this claim should be provided.

## Documentation

There is no clear documentation of the code's functionalities. I had to read the source code to understand some of its key functionalities. The authors need to generate documentation for the API that is not embedded in the source code.

- Statement of need: It is unclear why the authors provide a new solution to this problem. Do the authors benchmark their work against more established libraries such as MDAnalysis, MDTraj or TRAVIS?
- Installation instructions: The installation does not work as described by the authors, as mentioned before.
- Example usage: The examples do not work as described by the authors. I have to modify the provided commands to make them work. Additionally, it is unclear what results are obtained; it would be better if the examples were described in more detail. On the other hand, the execution time of the examples is considerably long. This is not explicitly mentioned. The first example took around 13 minutes using 8 cores. The second (`analyze_correlations`) did not work. The third (`analyze_descriptors`) yields an `AttributeError`. Finally, the script `run.sh` refers to a local path, leading to confusion.

## Software paper

The paper does not contain a statement of need. Additionally, it does not describe the current state of the art or how this new software can compare with well-established codes for analyzing MD trajectories.

- Summary: The authors must specify their target audience. If I understand correctly, their code can only be used to analyse trajectories from _ab-initio_ simulations, which excludes people who make classical simulations. Even among people who make  _ab initio_ simulations, it seems their code is targeted to be used with VASP, reducing the pool of possible users considerably.
- State-of-the-art: The authors must add appropriate citations to some basic concepts to develop the paper. For example, the definition of fast-ion conductors requires a reference to where it is taken. Additionally, citations to other codes that can be used for a similar task must be added if those exist.
- References: As mentioned, the authors must enrich their references to cover the state-of-the-art.

In addition, I have the following questions:

- The authors do not adequately specify the format of the MD trajectory that needs to be used in their code. This must be clarified because the code seems heavily dependent on only that specific format. Complementary, the specification of the trajectory format helps ensure the reliability of results obtained with other software.
- The authors mentioned that one of the code's functionalities is `analyze_descriptors`. I think a more detailed description of which descriptors are obtained and how they are obtained is required.
- From the technical side, the authors need to provide information about hyperparameters for the k-means algorithm or give a criterion, either empirically or physically based, for selecting those values that are known to be key to using the unsupervised method.
- In the same line as above, I wonder if the authors consider using other unsupervised machine learning algorithms (see for examples, section 5 of Chem. Rev. 2021, 121, 16, 9722–9758) to do the clustering. It is mentioned that it is possible to use spectral clustering. If that is the case, an explicit example should be added.
- Regarding the methods section, in equation 1, the authors mention that the ionic conductivity is a function of the number of spatial dimensions. I would like to know if the code allows the computation of the ionic conductivity in a specific dimension (i.e., in the x-axis only).
- Authors need to provide information about the equation used to obtain the diffusion coefficient; equation 2 seems to be the Einstein relationship, while equation 4 seems like a discretized version of the Green-Kubo expression. To do this, I encourage the authors to review the following reference: Living J. Comput. Mol. Sci., 2019, 1, 6324.
- The authors mention that they assumed that the non-diffusive particles have an isotropic behaviour in their fluctuations. What is the reason for that assumption? In which cases is this assumption fulfilled? Is the developed code still useful if the non-diffusive particles do not follow that assumption?
- The authors adopt an arbitrary threshold of 0.5 $\unicode{x212B}$. What is the justification for this definition? As they mentioned, the definition of this distance is critical for obtaining the k-centers. Then, what is the effect of this value? Could the author provide information on what happens if other threshold values are used?
- In the subsection regarding correlations between mobile ions, the procedure explained in the first paragraph can be illustrated as pseudocode to facilitate its comprehension.
- Regarding the examples, why do you always obtain one or two clusters? Is that what we should expect? Does that mean that only one diffusion path is observed in your simulations? What is the physical meaning of the clusters obtained by the code, if there is one?

## Suggested improvements

Besides the points mentioned before, I think the author could make the following improvements:
- Improve the code installation using a `setup.py` file or a `.toml` file. Also, a mention of specific requirements such as Python and package version will be very useful.
- Create detailed tutorials for using the code with the aid of either scripts or Jupyter notebooks that allow the user to understand what the code is doing.
- The authors must provide a better description of their code's capabilities. Judging from their software's reported use, I believe this code has multiple applications. However, it is unclear to me which capabilities and how they can be applied by the MD community.
- As mentioned before, documentation must be improved. I also suggest that the authors include a detailed section explaining the theory behind it. This should not be a problem because the authors explained it in their previous work. Last but not least, a section on the documentation on constructing the input files for the code must be added.
- I suggest the authors add more details about generating the MD Trajectories. As these simulations are very sensitive to different factors, the user must be alerted that their results will depend on those.

## Overall

I believe the presented work can be useful for analysing MD trajectories for ionic systems. However, I have several doubts about the software's functionality and applications; therefore, the paper requires major revision to have a clear view of them. Also, the state-of-the-art software must be improved, and suitable comparisons with other available software must be made. Finally, I failed to see how this work can be combined with existing frameworks as it seems to be specialized in analyzing VASP trajectories and is incompatible with classical MD. Likewise, I failed to see how this code systematically determines exact ionic migration paths and the level of coordination between mobile particles in diffusive events. To conclude, the paper and the code must go through major revision before considering its acceptance.

## Other questions

- In one of the previous works by the authors, the van Hove (VH) autocorrelation function is reported. Is it possible to obtain the VH function with the present code?
