### 2024-05-09 IonRepo/IonDiff#2 @yw-fang
[REVIEW comments] for the paper and code submitted to JOSS
Hello @CibranLopez and @IonRepo I'll append the comments in the coming days gradually on your manuscript and code submitted to Journal of Open Source Software https://github.com/openjournals/joss-reviews/issues/6616

Because the comments are appended now and then during conferences, I leave the information here so that you know if I have finished up the review:
- [ ] Is my review report complete?

The details comments are as follows:

The motivation described in the manuscript for implementing this code is clear.

## On manuscript
- [ ]  On line 32, I think it omitted "they'' between "valuable'' and "since''
