+++ docs/requirements.txt
@@ -1,0 +1,5 @@
+numpy
+scipy
+scikit-learn
+multiprocess
+seaborn
