+++ docs/conf.py
@@ -1,0 +1,242 @@
+"""
+Shared Sphinx configuration using sphinx-multiproject.
+
+To build each project, the ``PROJECT`` environment variable is used.
+
+.. code:: console
+
+   $ make html  # build default project
+   $ PROJECT=dev make html  # build the dev project
+
+for more information read https://sphinx-multiproject.readthedocs.io/.
+"""
+
+import os
+import sys
+
+from multiproject.utils import get_project
+
+sys.path.append(os.path.abspath("_ext"))
+extensions = [
+    "hoverxref.extension",
+    "multiproject",
+    "myst_parser",
+    "notfound.extension",
+    "sphinx_copybutton",
+    "sphinx_design",
+    "sphinx_tabs.tabs",
+    "sphinx-prompt",
+    "sphinx.ext.autodoc",
+    "sphinx.ext.autosectionlabel",
+    "sphinx.ext.extlinks",
+    "sphinx.ext.intersphinx",
+    "sphinxcontrib.httpdomain",
+    "sphinxcontrib.video",
+    "sphinxemoji.sphinxemoji",
+    "sphinxext.opengraph",
+]
+
+multiproject_projects = {
+    "user": {
+        "use_config_file": False,
+        "config": {
+            "project": "Read the Docs user documentation",
+            "html_title": "Read the Docs user documentation",
+        },
+    },
+    "dev": {
+        "use_config_file": False,
+        "config": {
+            "project": "Read the Docs developer documentation",
+            "html_title": "Read the Docs developer documentation",
+        },
+    },
+}
+
+docset = get_project(multiproject_projects)
+
+ogp_site_name = "Read the Docs Documentation"
+ogp_use_first_image = True  # https://github.com/readthedocs/blog/pull/118
+ogp_image = "https://docs.readthedocs.io/en/latest/_static/img/logo-opengraph.png"
+# Inspired by https://github.com/executablebooks/MyST-Parser/pull/404/
+ogp_custom_meta_tags = [
+    '<meta name="twitter:card" content="summary_large_image" />',
+]
+ogp_enable_meta_description = True
+ogp_description_length = 300
+
+templates_path = ["_templates"]
+
+# This may be elevated as a general issue for documentation and behavioral
+# change to the Sphinx build:
+# This will ensure that we use the correctly set environment for canonical URLs
+# Old Read the Docs injections makes it point only to the default version,
+# for instance /en/stable/
+html_baseurl = os.environ.get("READTHEDOCS_CANONICAL_URL", "/")
+
+master_doc = "index"
+copyright = "Read the Docs, Inc & contributors"
+version = "10.27.0"
+release = version
+exclude_patterns = ["_build", "shared", "_includes"]
+default_role = "obj"
+intersphinx_cache_limit = 14  # cache for 2 weeks
+intersphinx_timeout = 3  # 3 seconds timeout
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3.10/", None),
+    "django": (
+        "https://docs.djangoproject.com/en/stable/",
+        "https://docs.djangoproject.com/en/stable/_objects/",
+    ),
+    "sphinx": ("https://www.sphinx-doc.org/en/master/", None),
+    "pip": ("https://pip.pypa.io/en/stable/", None),
+    "nbsphinx": ("https://nbsphinx.readthedocs.io/en/latest/", None),
+    "myst-nb": ("https://myst-nb.readthedocs.io/en/stable/", None),
+    "ipywidgets": ("https://ipywidgets.readthedocs.io/en/stable/", None),
[…]
