+++ docs/conf.py
@@ -1,242 +1,66 @@
-"""
-Shared Sphinx configuration using sphinx-multiproject.
+# Configuration file for the Sphinx documentation builder.
+#
+# This file only contains a selection of the most common options. For a full
+# list see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
 
-To build each project, the ``PROJECT`` environment variable is used.
+# -- Path setup --------------------------------------------------------------
 
-.. code:: console
+# If extensions (or modules to document with autodoc) are in another directory,
+# add these directories to sys.path here. If the directory is relative to the
+# documentation root, use os.path.abspath to make it absolute, like shown here.
+#
+# import os
+# import sys
+# sys.path.insert(0, os.path.abspath('.'))
 
-   $ make html  # build default project
-   $ PROJECT=dev make html  # build the dev project
 
-for more information read https://sphinx-multiproject.readthedocs.io/.
-"""
+# -- Project information -----------------------------------------------------
 
-import os
-import sys
+project = "IonDiff"
+copyright = "2024, Cibrán López Álvarez"
+author = "Cibrán López Álvarez"
 
-from multiproject.utils import get_project
 
-sys.path.append(os.path.abspath("_ext"))
+# -- General configuration ---------------------------------------------------
+
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
+# ones.
 extensions = [
-    "hoverxref.extension",
-    "multiproject",
-    "myst_parser",
-    "notfound.extension",
-    "sphinx_copybutton",
-    "sphinx_design",
-    "sphinx_tabs.tabs",
-    "sphinx-prompt",
     "sphinx.ext.autodoc",
-    "sphinx.ext.autosectionlabel",
-    "sphinx.ext.extlinks",
-    "sphinx.ext.intersphinx",
-    "sphinxcontrib.httpdomain",
-    "sphinxcontrib.video",
-    "sphinxemoji.sphinxemoji",
-    "sphinxext.opengraph",
+    "sphinx.ext.napoleon",
+    "sphinx_rtd_theme",
+    "myst_parser",
 ]
+source_suffix = [".rst", ".md"]
 
-multiproject_projects = {
-    "user": {
-        "use_config_file": False,
-        "config": {
-            "project": "Read the Docs user documentation",
-            "html_title": "Read the Docs user documentation",
-        },
-    },
-    "dev": {
-        "use_config_file": False,
-        "config": {
-            "project": "Read the Docs developer documentation",
-            "html_title": "Read the Docs developer documentation",
-        },
-    },
-}
-
-docset = get_project(multiproject_projects)
-
-ogp_site_name = "Read the Docs Documentation"
-ogp_use_first_image = True  # https://github.com/readthedocs/blog/pull/118
-ogp_image = "https://docs.readthedocs.io/en/latest/_static/img/logo-opengraph.png"
-# Inspired by https://github.com/executablebooks/MyST-Parser/pull/404/
-ogp_custom_meta_tags = [
-    '<meta name="twitter:card" content="summary_large_image" />',
-]
-ogp_enable_meta_description = True
-ogp_description_length = 300
-
+# Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
-# This may be elevated as a general issue for documentation and behavioral
-# change to the Sphinx build:
[…]
