+++ docs/conf.py
@@ -48,15 +48,14 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-#
-html_theme = "sphinx_rtd_theme"
+html_theme = "sphinx_book_theme"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-html_favicon = "favicon.png"
-html_logo = "logo.png"
+html_favicon = "logo.svg"
+html_logo = "logo.svg"
 html_theme_options = {
     "logo_only": True,
 }
