+++ setup.py
@@ -1,6 +1,6 @@
 from distutils.core import setup
 
-with open('README.rst') as f:
+with open('README.md') as f:
     readme = f.read()
 
 with open('LICENSE') as f:
