+++ setup.py
@@ -1,6 +1,6 @@
 from distutils.core import setup
 
-with open('README.md') as f:
+with open('docs/index.rst') as f:
     readme = f.read()
 
 with open('LICENSE') as f:
