+++ docs/index.rst
@@ -1,7 +1,7 @@
 IonDiff
 =======
 
-.. image:: logo.svg
+.. image:: https://github.com/IonRepo/IonDiff/blob/0.3/docs/logo.svg
    :width: 20%
 
 Despite playing a central role in the design of high performance solid-state electrolytes, little is known about the processes governing ionic diffusion in these materials and the spatio-temporal correlations acting on migrating particles. Computer simulations can reproduce the trajectories of individual diffusing ions in real time with extraordinary accuracy, thus providing incredibly valuable atomistic data that in practice cannot be resolved by experiments.
