### 2024-06-07 9f68967
Update analyze_descriptors.py
