+++ settings.py
@@ -1,5 +1,0 @@
-from pathlib import Path
-
-BASE_DIR = Path(__file__).parent
-
-LOGS_PATH   = BASE_DIR / 'logs'
