+++ build/lib/IonDiff/identify_diffusion.py
@@ -1,0 +1,244 @@
+#!/usr/bin/env python
+import numpy             as np
+import matplotlib.pyplot as plt
+import seaborn           as sns
+import sys
+import os
+
+from sklearn.cluster import KMeans, SpectralClustering
+from sklearn.metrics import silhouette_score
+
+from IonDiff import common_library as CL
+
+sns.set_theme()
+
+"""Definition of the class to extract the diffusion information. Only VASP simulations are considered as yet.
+"""
+
+# Defining the basic parameters for k-means and spectral clustering algorithms
+kmeans_kwargs   = dict(init='random', n_init=10, max_iter=300, tol=1e-04,                          random_state=0)
+spectral_kwargs = dict(affinity='nearest_neighbors', n_neighbors=1000, assign_labels='cluster_qr', random_state=0)
+
+# Defining the class
+class xdatcar:
+    """Python Class for loading information from VASP simulations and identifying diffusions.
+
+    Attributes:
+        kmeans_kwargs   (dict):  Parameters for K-means clustering algorithm.
+        spectral_kwargs (dict):  Parameters for Spectral clustering algorithm.
+        time_step       (float): Time step between consecutive XDATCAR configurations.
+        n_ions          (int):   Number of ions in the simulation.
+
+    Methods:
+        __init__(self, args):
+            Initializes the XDATCAR class.
+        read_simulation(self, args):
+            Reads VASP XDATCAR files.
+        calculate_silhouette(self, coordinates, method, n_attempts, silhouette_thd):
+            Calculates silhouette scores for different numbers of clusters and selects the optimal number.
+        calculate_clusters(self, coordinates, n_clusters, method, distance_thd):
+            Calculates clusters and related information based on the chosen method and number of clusters.
+        get_diffusion(self, args):
+            Obtains diffusion information from the simulation data.
+    """
+
+    def __init__(self, args):
+        """Initialize the XDATCAR class.
+
+        Args:
+            args: Command line arguments containing MD_path and other parameters.
+
+        Raises:
+            exit: If required files are missing.
+        """
+        
+        # Loading intervals step and number of simulation steps between records
+        delta_t, n_steps = CL.read_INCAR(args.MD_path)
+        
+        # Time step between consecutive XDATCAR configurations
+        self.time_step = n_steps * delta_t
+        
+        # Reading the simulation data
+        try:
+            self.read_simulation(args)
+        except:
+            sys.exit('Error: simulation not found')
+
+
+    def read_simulation(self, args):
+        """Reads VASP XDATCAR files.
+        """
+        
+        cell, self.n_ions, _, _, position = CL.read_XDATCAR(args.MD_path)
+        
+        self.positionC = np.zeros_like(position)
+        self.n_iter    = position.shape[0]
+        
+        # Getting the variation in positions and applying periodic boundary condition
+        dpos = np.diff(position, axis=0)
+        dpos[dpos >  0.5] -= 1.0
+        dpos[dpos < -0.5] += 1.0
+        
+        # Getting the positions and variations in cell units
+        for i in range(self.n_iter-1):
+            self.positionC[i] = np.dot(position[i], cell)
+            dpos[i]           = np.dot(dpos[i],     cell)
+        
+        self.positionC[-1] = np.dot(position[-1], cell)
+        
+        # Defining the attribute of window=1 variation in position and velocity
+        self.velocity = dpos / self.time_step
+        self.d1pos    = dpos
+    
+    
+    def calculate_silhouette(self, coordinates, method, n_attempts, silhouette_thd):
+        """Calculates silhouette scores for different numbers of clusters and selects the optimal number.
+
+        Args:
+            coordinates    (numpy.ndarray): Input array of coordinates.
[…]
