+++ build/lib/IonDiff/analyze_descriptors.py
@@ -1,0 +1,235 @@
+#!/usr/bin/env python
+import numpy as np
+import json
+import sys
+
+import IonDiff.common_library as CL
+
+"""Definition of the class to analyse correlations among descriptors of diffusive paths.
+   A database with more than one simulation must be provided, as well as paths to every
+   interesting DIFFUSION file, extracted from the simulation.
+"""
+
+# Defining the class
+class descriptors:
+    """Python Class for loading information from VASP simulations
+       and analyzing correlations among descriptors of diffusive paths.
+
+    Methods:
+        __init__(self, args):
+            Initializes the Descriptors class.
+    """
+
+    def __init__(self, args):
+        """Initialize the Database class.
+
+        Args:
+            args: Command line arguments containing MD_path and other parameters.
+
+        Raises:
+            exit: If required files are missing.
+        """
+        
+        # Loading the data
+        coordinates, self.hoppings, self.cell, self.compounds, self.concentration = CL.load_data(args.MD_path)
+        (self.n_conf, self.n_particles, _) = np.shape(coordinates)
+
+        # Getting diffusion coordinates
+        _, nan_hoppings = CL.get_expanded_hoppings(self.n_conf, self.n_particles, self.concentration, self.compounds,
+                                                   self.hoppings, method='original')
+        nan_hoppings[nan_hoppings == 0] = np.NaN
+
+        self.coordinates = np.stack([nan_hoppings, nan_hoppings, nan_hoppings], axis=2) * coordinates
+
+        # Expanding the hoppings avoiding non-diffusive particles
+        _, self.expanded_hoppings = CL.get_expanded_hoppings(self.n_conf, self.n_particles, self.concentration,
+                                                             self.compounds, self.hoppings, method='separate')
+    
+
+    def time_until_diffusion(self, index=0, outer=None):
+        """
+        Calculates an array with the number of steps until the start of the first diffusion of a particle.
+        Expanded matrix in 'separate' or 'cleaned' format is mandatory.
+        We seek the first one for each particle and append it to the list (indexes are in agreement with 'key').
+        Index = {0, -1} variable allows obtaining the first or last diffusive position of a diffusion track.
+        Calculate Time Until Diffusion
+
+        Args:
+            expanded_matrix (numpy.ndarray): Expanded matrix of diffusive/non-diffusive processes.
+            index           (int, optional): Index to obtain the first or last diffusive position (default is 0).
+            outer           (optional):      Value for handling NaN values (default is None).
+
+        Returns:
+            numpy.ndarray: Array with time until diffusion for each particle.
+        """
+        n_particles = np.shape(self.expanded_matrix)[1]
+        initial_times = np.zeros(n_particles)
+        for i in range(n_particles):
+            particle_track = self.expanded_matrix[:, i]
+            if outer is None:
+                initial_times[i] = np.where(particle_track)[0][index]
+            elif outer == 'nan':
+                initial_times[i] = np.where(~np.isnan(particle_track))[0][index]
+        return initial_times
+    
+    
+    def duration_of_diffusion(self):
+        """
+        Calculates an array with the number of steps of diffusion of a particle.
+        Expanded hoppings in 'separate' format is mandatory.
+        We seek the first one for each particle and append it to the list (indexes are in agreement with 'key').
+        Calculate Duration of Diffusion
+
+        Args:
+            expanded_hoppings (numpy.ndarray): Expanded hopping data.
+
+        Returns:
+            numpy.ndarray: Array with the duration of diffusion for each particle.
+        """
+
+        diffusion_duration = time_until_diffusion(self.expanded_hoppings, index=-1) - time_until_diffusion(self.expanded_hoppings, index=0)
[…]

[…]
