### 2024-06-19 #6616 @Claudiator7
Dear Rachel,

We are working on the revision of our article and IonDiff software by following all the Reviewers recommendations. There were many points to consider, specially from Referee 1, so that this is taking us more time than what we initially expected. We think that we will be able to come back to you with the revised article and improved software in a couple of weeks or so. Sorry about our delay.

Regards,
Claudio
