+++ paper/paper.bib
@@ -85,3 +85,23 @@
   	year = {2023},
 	doi = {10.1038/s41524-023-00996-8}
 }
+
+@article{Hull2004,
+  	author = {{Hull}, S.},
+  	title = {{Superionics: crystal structures and conduction processes}},
+	journal = {Rep. Prog. Phys.},
+        volume = {67},
+	pages = {1233},
+  	year = {2004},
+	doi = {10.1088/0034-4885/67/7/R05}
+}
+
+@article{Sagotra2017,
+  	author = {{Sagotra}, A. K. and {Cazorla}, C.},
+  	title = {{Stress-mediated enhancement of ionic conductivity in fast-ion conductors}},
+	journal = {ACS Appl. Mater. Interfaces.},
+        volume = {9},
+	pages = {38773},
+  	year = {2017},
+	doi = {10.1021/acsami.7b11687}
+}
