### 2024-06-30 f8a0dc2
Update ionic-diffusion coefficient computation
