+++ IonDiff/identify_diffusion.py
@@ -3,7 +3,6 @@ import numpy             as np
 import matplotlib.pyplot as plt
 import seaborn           as sns
 import sys
-import os
 
 from IonDiff import common_library as CL
 
@@ -46,10 +45,10 @@ class xdatcar:
         """
         
         # Loading intervals step and number of simulation steps between records
-        delta_t, n_steps = CL.read_INCAR(args.MD_path)
+        self.delta_t, self.n_steps = CL.read_INCAR(args.MD_path)
         
         # Time step between consecutive XDATCAR configurations
-        self.time_step = n_steps * delta_t
+        self.time_step = self.n_steps * self.delta_t
         
         # Reading the simulation data
         try:
@@ -61,8 +60,8 @@ class xdatcar:
     def read_simulation(self, args):
         """Reads VASP XDATCAR files.
         """
-        
-        cell, self.n_ions, _, _, position = CL.read_XDATCAR(args.MD_path)
+
+        cell, self.n_ions, self.compounds, self.concentration, position = CL.read_XDATCAR(args.MD_path)
         
         self.positionC = np.zeros_like(position)
         self.n_iter    = position.shape[0]
@@ -76,13 +75,17 @@ class xdatcar:
         for i in range(self.n_iter-1):
             self.positionC[i] = np.dot(position[i], cell)
             dpos[i]           = np.dot(dpos[i],     cell)
-        
+
         self.positionC[-1] = np.dot(position[-1], cell)
+        self.dpos          = dpos
+
+        expanded_dimensions        = np.expand_dims(self.positionC[0], 0)
+        self.cartesian_coordinates = np.concatenate([expanded_dimensions, self.dpos], axis=0)
+        self.cartesian_coordinates = np.cumsum(self.cartesian_coordinates, axis=0)
         
         # Defining the attribute of window=1 variation in position and velocity
-        self.velocity = dpos / self.time_step
-        self.d1pos    = dpos
-    
+        self.velocity = self.dpos / self.time_step
+
     
     def get_diffusion(self, args):
         """Obtains diffusion information from the simulation data.
@@ -95,10 +98,6 @@ class xdatcar:
         """
         
         hoppings = []
-        
-        full_coordinates = np.concatenate([np.expand_dims(self.positionC[0], 0), self.d1pos], axis=0)
-        full_coordinates = np.cumsum(full_coordinates, axis=0)
-        
         if args.make_plot:
             fig = plt.figure(figsize=(5, 5))
             ax = plt.axes(projection='3d')
@@ -107,7 +106,7 @@ class xdatcar:
             ax.set_zlabel('z')
         
         for particle in range(self.n_ions):
-            coordinates = full_coordinates[:, particle]
+            coordinates = self.cartesian_coordinates[:, particle]
             
             # Recommended number of clusters
             n_clusters = CL.calculate_silhouette(coordinates,
@@ -165,3 +164,144 @@ class xdatcar:
                     plt.legend(loc='best')
                 plt.show()
         return hoppings
+
+    def get_diffusion_coefficient(self, args, initial_ratio=0.4, ending_ratio=1, contributions='full', n_delta_t=None, axis=None):
+        """Calculates the diffusion coefficient from atomic trajectories using Mean Square Displacement (MSD) method.
+
+            Args:
+                args:                            Arguments containing parameters like `diffusion_particle` and `make_plot`.
+                initial_ratio (float, optional): Ratio of the initial point for fitting the MSD curve. Defaults to 0.4.
+                ending_ratio  (float, optional): Ratio of the ending point for fitting the MSD curve. Defaults to -1.
+                contributions (str, optional):   Specifies which contributions to consider ('full', 'self', 'distinct'). Defaults to 'full'.
[…]
