### 2024-07-01 #6616 @Claudiator7
Dear Rachel and Reviewers,

Thanks a lot for your patience and for very careful and comprehensive reading of our work and for their very useful and insightful comments that have allowed us to significantly improve the readability of our article and the user-friendliness of our IonDiff software.

Please, find attached a point-by-point response to all the Reviewers' comments and criticisms as well as a copy of the revised manuscript.

We trust that the revised version of our article and improved software meets the standards for publication in your esteemed journal.

[response-reviewer-LIVazquezS.pdf](https://github.com/user-attachments/files/16051967/response-reviewer-LIVazquezS.pdf)
[response-reviewer-yw-fang.pdf](https://github.com/user-attachments/files/16051968/response-reviewer-yw-fang.pdf)
[paper-rev.pdf](https://github.com/user-attachments/files/16051970/paper-rev.pdf)

Best wishes,
Claudio Cazorla
