### 2024-07-01 #6616 @Claudiator7
(Reproduced from GitHub thread)

Dear Rachel and Reviewers,

Thanks a lot for your patience and for very careful and comprehensive
reading of our work and for their very useful and insightful comments that
have allowed us to significantly improve the readability of our article and
the user-friendliness of our IonDiff software.

Please, find attached a point-by-point response to all the Reviewers'
comments and criticisms as well as a copy of the revised manuscript.

We trust that the revised version of our article and improved software
meets the standards for publication in your esteemed journal.

Best wishes,
Claudio Cazorla
======================================
Dr. Claudio Cazorla
"Ramón y Cajal" Researcher
Department of Physics
Universitat Politècnica de Catalunya (UPC)
C/ Jordi Girona 1-3, Campus Nord, B4-203
08034 Barcelona (Spain)
telephone: +34 93 4017066
email: ***@***.***
website: www.claudiocazorla-physics.com

Missatge de Rachel Kurchin ***@***.***> del dia dt., 18 de
juny 2024 a les 22:52:

> 🔔 Hi @Claudiator7 <https://github.com/Claudiator7>, just checking in
> here!
>
