+++ README.md
@@ -18,6 +18,14 @@ IonDiff can be installed from PyPI:
 pip3 install IonDiff
 ```
 
+or  directly from:
+
+```bash
+git clone https://github.com/IonRepo/IonDiff.git
+cd IonDiff
+pip3 install .
+```
+
 or used directly from source without explicit installation:
 
 ```bash
