+++ logs/README.md
@@ -1,0 +1,3 @@
+# Logs
+
+This folder automatically stores information about different runs of the code.
