+++ README.md
@@ -45,10 +45,10 @@ python3 cli.py identify_diffusion --MD_path examples
 To analyze temporal correlations among the diffusions of different simulations, from the IonDiff folder run:
 
 ```bash
-python3 cli.py analyze_correlations
+python3 cli.py analyze_correlations --MD_path examples
 ```
 
-and to extract atomistic descriptors from the simulation and diffusion events run:
+and to extract atomistic descriptors from the simulations and diffusion events run:
 
 ```bash
 python3 cli.py analyze_descriptors
