+++ docs/index.rst
@@ -54,6 +54,13 @@ An *ab initio* MD simulation based on density functional theory of non-stoichiom
 - `examples/XDATCAR <https://github.com/IonRepo/IonDiff/tree/main/examples/XDATCAR>`_: Concatenation of all simulated configurations (recorded each **NBLOCK** simulation steps).
 - `examples/README.md <https://github.com/IonRepo/IonDiff/tree/main/examples/README.md>`_: More specific information regarding these files.
 
+Input trajectories
+-------
+
+The IonDiff code can be perfectly used by any scientist performing either classical molecular dynamics simulations (classical MD) or ab initio molecular dynamics simulations (AIMD). In both types of simulations, atomic trajectories are generated and this is the main input information that the IonDiff code necessitates to perform its correlation and ionic hopping analysis. In other words, the IonDiff analysis does not depend on how the atomic forces are calculated in the undertaken molecular dynamics simulations, whether these are obtained through classical force fields or quantum mechanical methods (e.g., density functional theory). As far as the output trajectory files generated by any classical MD code can be converted to the output trajectory file format of VASP, the IonDiff code can be purposely employed.
+
+There is already a myriad of open-source codes and scripts that can be used for this end, namely, to convert a trajectory file generated by a classical MD code to the VASP format, like, for instance, `LAVA <https://github.com/lanl/LAVA>`_ or `xfroggie <https://www.xfroggie.com/index.cgi/strconv>`_ (to convert from LAMMPS or GROMACS format to VASP format, respectively).
+
 Authors
 -------
 
