### 2024-07-09 #6616 @IonRepo
Dear Reviewer,

We thank the Reviewer for their insightful comments and thorough review, as
well as their nice feedback regarding our improvements in both text and
package. We agree on the complexity of the mentioned task for non
experienced used, which can be challengeing and discouraging. Following the
reviewer's advice, we added a section explaining how to use the
trajectories generated with different codes as input to IonDiff in the
README (https://github.com/IonRepo/IonDiff/blob/main/README.md) and the
documentation (https://iondiff.readthedocs.io/en/latest/). As well, the
hyperlik to the documentation was not displaying correctly in the README
file, as pointed out, which has been now fixed. Finally, we thank the
reviewer for detecting that the tuple from line 77 in
analyze_correlations.py could be more efficiently unpacked, which has been
now improved.

Best regards,
Cibrán.

Missatge de Luis Itza Vazquez-Salazar ***@***.***> del dia
dl., 8 de jul. 2024 a les 16:11:

> Dear Authors,
>
> Thank you very much for your detailed response to the points raised before
> and for solving most of them.
>
> The paper has improved considerably. Nevertheless, I still have three
> points that require the authors' attention.
>
>    1. In their reply, the authors mention mention:
>    "*There seems to be a misunderstanding on this Reviewer’s point. The
>    IonDiff code can be perfectly used by any scientist performing either
>    classical molecular dynamics simulations (classical MD) or ab initio
>    molecular dynamics simulations (AIMD). In both types of simulations, atomic
>    trajectories are generated and this is the main input information that the
>    IonDiff code necessitates to perform its correlation and ionic hopping
>    analysis. In other words, the IonDiff analysis does not depend on how the
>    atomic forces are calculated in the undertaken molecular dynamics
>    simulations, whether these are obtained through classical force fields or
>    quantum mechanical methods (e.g., density functional theory). As far as the
>    output trajectory files generated by any classical MD code can be converted
>    to the output trajectory file format of VASP, which is very simple and
>    intuitive, the IonDiff code can be purposely employed. There is already a
>    myriad of open-source codes and scripts that can be used for this end,
>    namely, to convert a trajectory file generated by a classical MD code to
>    the VASP format, like, for instance, LAVA (https://github.com/lanl/LAVA
>    <https://github.com/lanl/LAVA>, to convert from LAMMPS format to VASP
>    format) and the online website https://www.xfroggie.com/index.cgi/strconv
>    <https://www.xfroggie.com/index.cgi/strconv> (to convert from GROMACS
>    format to VASP format). Therefore, there is not any factual limitation on
>    the number of potential users of the IonDiff code. Likewise, there is
>    neither a pressing need to replicate already existing codes and scripts
>    able to perform MD files format conversions.*"
>
> The authors must explicitly mention this in the documentation. Although
> this might be obvious for the experienced user, it might be a big headache
> for an inexpert user because it is not completely obvious how to convert
> files generated with one code to the required format by IonDiff.
>
>    2.
>
>    The authors mention "*Generation of documentation in ReadTheDocs*." I
>    have not found the link to this documentation, so I would like to kindly
>    ask the authors to add it to the README file.
>    3.
>
>    While running the examples provided by the authors, I found a minor
>    bug in the first example with analyze_correlations.py in line 77 with
>    the number of arguments to unpack.
>
> Finally, I would like to acknowledge the author's effort to improve their
> code and the supporting manuscript.
>
