### 2024-07-11 #6616 @LIVazquezS
Dear Authors,

Thanks for your efforts to improve the code. I am fine with the manuscript associated with the code. Nevertheless, I still have two problems with the code.

1. The bug in the first example with `analyze_correlations.py` in line 77 with the number of arguments to unpack. I am using python 3.8.19. Adjuncted a  screenshot of the mentioned problem.

2. Thanks for adding the link to the read-the-docs page with the documentation; it works fine. Nevertheless, to fullfil the requirements of JOSS, you need to add API level documentation. This should not be a problem as all functions in the code are adequately documented, so any automatic API generator like sphinx [https://www.sphinx-doc.org/en/master/man/sphinx-apidoc.html] can do it. This is a minor point, but it will improve the user experience as the arguments of each function can be checked from the documentation instead of reading the source code.

Again, thanks for your effort. Once those two points are sorted, I will gladly recommend accepting your work.
