+++ source/index.rst
@@ -1,0 +1,20 @@
+.. IonDiff documentation master file, created by
+   sphinx-quickstart on Thu Jul 11 17:32:44 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to IonDiff's documentation!
+===================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
