+++ build/lib/IonDiff/analyze_correlations.py
@@ -1,240 +1,0 @@
-#!/usr/bin/env python
-import numpy             as np
-import matplotlib.pyplot as plt
-import seaborn           as sns
-import multiprocess      as mp
-import sys
-import os
-
-import IonDiff.common_library as CL
-
-from scipy.optimize import curve_fit
-
-sns.set_theme()
-
-"""Definition of the class to analyse correlations among diffusive paths. A database with more than one
-   simulation must be provided, as well as paths to every interesting DIFFUSION file, extracted from the simulation.
-"""
-
-fontsize = 10
-dpi      = 100
-
-# Defining the class
-class database:
-    """Python Class for loading information from VASP simulations and analyzing correlations among diffusive paths.
-
-    Attributes:
-        fontsize (int): Font size for plot labels and ticks.
-        dpi      (int): Dots per inch for saving plots.
-    
-    Methods:
-        __init__(self, args):
-            Initializes the Database class.
-        exp_function(self, x, A, B, C):
-            Defines a decreasing and always positive exponential function.
-        parallel_calculation(self, element, args):
-            Auxiliar function to parallelize calculations.
-    """
-
-    def __init__(self, args):
-        """Initialize the Database class.
-
-        Args:
-            args: Command line arguments containing MD_path and other parameters.
-
-        Raises:
-            exit: If required files are missing.
-        """
-        
-        correlations_matrix_path = f'{args.MD_path}/temp_matrix'
-        DIFFUSION_paths          = f'{args.MD_path}/DIFFUSION_paths'
-        
-        if os.path.exists(correlations_matrix_path):
-            print('As temp_matrix already exists, its information is being loaded.')
-            
-            with open(correlations_matrix_path, 'r') as file:
-                lines = file.readlines()
-            
-            n = int(len(lines) / 3)
-            length = len(lines[0].split())
-
-            corr_matrix = np.zeros((n, length))
-            temp_matrix = np.zeros((n, length))
-            fami_matrix = np.zeros((n, length), dtype=object)
-            for i in range(n):
-                corr_matrix[i] = np.array(lines[i].split(), dtype=float)
-                temp_matrix[i] = np.array(lines[i+n].split(), dtype=float)
-                fami_matrix[i] = np.array(lines[i+2*n].split(), dtype=object)
-        
-        elif os.path.exists(DIFFUSION_paths):
-            print('As temp_matrix does not exist, it is being computed.')
-            
-            with open(DIFFUSION_paths, 'r') as file:
-                paths_to_DIFFUSION = file.readlines()
-            
-            # The computation of correlations for each element are parallelized
-            pool    = mp.Pool(mp.cpu_count())  # Number of CPUs in the PC
-            metrics = [pool.apply(self.parallel_calculation, (element, args,)) for element in paths_to_DIFFUSION]
-            pool.close()
-            
-            n_events = 0
-            for lines in metrics:
-                line = lines[0]
-                if len(line) > n_events:
-                    n_events = len(line)
-
-            n_simulations = len(metrics)
-
-            corr_matrix = np.zeros((n_simulations, n_events))
-            temp_matrix = np.ones((n_simulations, n_events)) * np.NaN
-            fami_matrix = np.zeros((n_simulations, n_events), object)
-
-            for i in range(n_simulations):
-                corr_temporal = np.array(metrics[i][0])
-                temp_temporal = np.array(metrics[i][1])
-                fami_temporal = np.array(metrics[i][2])
-                
[…]
