+++ build/lib/IonDiff/common_library.py
@@ -1,529 +1,0 @@
-import numpy             as np
-import matplotlib.pyplot as plt
-import pandas            as pd
-import re
-import os
-import sys
-
-from sklearn.cluster import KMeans, SpectralClustering
-from sklearn.metrics import silhouette_score
-from scipy.ndimage   import gaussian_filter1d
-from scipy.stats     import pearsonr, spearmanr
-
-# Defining the basic parameters for k-means and spectral clustering algorithms
-kmeans_kwargs   = dict(init='random', n_init=10, max_iter=300, tol=1e-04,                          random_state=0)
-spectral_kwargs = dict(affinity='nearest_neighbors', n_neighbors=1000, assign_labels='cluster_qr', random_state=0)
-
-"""Set of common functions for many libraries, with a general purpose.
-"""
-
-# Defining the information lines of the file
-scale_line         = 1  # Line for the scale of the simulation box
-s_cell_line        = 2  # Start of the definition of the simulation box
-e_cell_line        = 4  # End of the definition of the simulation box
-name_line          = 5  # Composition of the compound
-concentration_line = 6  # Concentration of the compound
-x_line             = 7  # Start of the simulation data
-
-
-def obtain_diffusive_information(composition, concentration, DiffTypeName=None):
-    """Gets the diffusive and non-diffusive elements.
-    
-    Args:
-        composition   (list):           List of element symbols in the compound.
-        concentration (list):           List of element concentrations.
-        DiffTypeName  (list, optional): List of diffusive element symbols. Defaults to None.
-    
-    Returns:
-        tuple: Tuple containing DiffTypeName (list), NonDiffTypeName (list), and diffusion_information (list).
-    
-    Example:
-        Obtain the diffusive information from a given composition and concentration.
-    
-        Usage:
-            obtain_diffusive_information(['Li', 'O'], [1, 2])
-    
-    Note:
-        If DiffTypeName is not provided, the function automatically selects diffusive elements based on an implicit order of preference.
-    """
-
-    if not DiffTypeName:
-        # Getting the diffusive element following the implicit order of preference
-        for diff_component in ['Li', 'Na', 'Ag', 'Cu', 'Cl', 'I', 'Br', 'F', 'O']:
-            if diff_component in composition:
-                DiffTypeName = [diff_component]
-                break
-
-    NonDiffTypeName = composition.copy()
-    for diff_element in DiffTypeName:
-        NonDiffTypeName.remove(diff_element)
-
-    # Getting the positions of the diffusive elements regarding the XDATCAR file
-    diffusion_information = []
-    for DiffTypeName_value in DiffTypeName:
-        for TypeName_index in range(len(composition)):
-            if composition[TypeName_index] == DiffTypeName_value:
-                diffusion_information.append([TypeName_index, np.sum(concentration[:TypeName_index]),
-                                              concentration[TypeName_index]])
-    return DiffTypeName, NonDiffTypeName, diffusion_information
-
-
-def obtain_diffusive_family(element):
-    """Returns the family of an element using the obtain_diffusive_information function.
-    
-    Args:
-        element (str): Element symbol.
-    
-    Returns:
-        str: Diffusive family of the element.
-    
-    Example:
-        Obtain the diffusive family of a given element.
-    
-        Usage:
-            obtain_diffusive_family('Li')
-    
-    Note:
-        The function automatically detects components from upper cases and numbers.
-    """
-
-    composition = []
[…]
