+++ docs/conf.py
@@ -14,15 +14,18 @@
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
 
-
 # -- Project information -----------------------------------------------------
 
-project = "IonDiff"
-copyright = "2024, Cibrán López Álvarez"
-author = "Cibrán López Álvarez"
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
+project = 'IonDiff'
+copyright = '2024, Cibrán López Álvarez'
+author = 'Cibrán López Álvarez'
+release = '1.4'
 
 # -- General configuration ---------------------------------------------------
+
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -41,12 +44,12 @@ extensions = [
 source_suffix = [".rst", ".md"]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
+templates_path = ['_templates']
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
 
 # -- Options for HTML output -------------------------------------------------
