+++ docs/source/IonDiff.rst
@@ -1,45 +1,0 @@
-IonDiff package
-===============
-
-Submodules
-----------
-
-IonDiff.analyze\_correlations module
-------------------------------------
-
-.. automodule:: IonDiff.analyze_correlations
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-IonDiff.analyze\_descriptors module
------------------------------------
-
-.. automodule:: IonDiff.analyze_descriptors
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-IonDiff.common\_library module
-------------------------------
-
-.. automodule:: IonDiff.common_library
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-IonDiff.identify\_diffusion module
-----------------------------------
-
-.. automodule:: IonDiff.identify_diffusion
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: IonDiff
-   :members:
-   :undoc-members:
-   :show-inheritance:
