+++ docs/source/modules.rst
@@ -1,7 +1,0 @@
-IonDiff
-=======
-
-.. toctree::
-   :maxdepth: 4
-
-   IonDiff
