+++ docs/index.rst
@@ -55,7 +55,7 @@ An *ab initio* MD simulation based on density functional theory of non-stoichiom
 - `README.md <https://github.com/IonRepo/IonDiff/tree/main/examples/README.md>`_: More specific information regarding these files.
 
 Input trajectories
--------
+------------------
 
 The IonDiff code can be perfectly used by any scientist performing either classical molecular dynamics simulations (classical MD) or ab initio molecular dynamics simulations (AIMD). In both types of simulations, atomic trajectories are generated and this is the main input information that the IonDiff code necessitates to perform its correlation and ionic hopping analysis. In other words, the IonDiff analysis does not depend on how the atomic forces are calculated in the undertaken molecular dynamics simulations, whether these are obtained through classical force fields or quantum mechanical methods (e.g., density functional theory). As far as the output trajectory files generated by any classical MD code can be converted to the output trajectory file format of VASP, the IonDiff code can be purposely employed.
 
@@ -68,8 +68,8 @@ API Documentation
    :maxdepth: 2
    :caption: Contents:
 
-   _build/html/source/IonDiff
-   _build/html/source/modules
+   _build/html/_sources/source/IonDiff
+   _build/html/_sources/source/modules
 
 Authors
 -------
