+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+IonDiff
+=======
+
+.. toctree::
+   :maxdepth: 4
+
+   IonDiff
