+++ docs/index.rst
@@ -65,7 +65,7 @@ API Documentation
 -----------------
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 4
    :caption: Contents:
 
    source/IonDiff
