+++ docs/index.rst
@@ -65,7 +65,7 @@ API Documentation
 -----------------
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 2
    :caption: Contents:
 
    source/IonDiff
