+++ docs/source/modules.rst
@@ -5,3 +5,7 @@ IonDiff
    :maxdepth: 4
 
    IonDiff
+   cli
+   settings
+   setup
+   tests
