+++ docs/index.rst
@@ -65,11 +65,15 @@ API Documentation
 -----------------
 
 .. toctree::
-   :maxdepth: 4
+   :maxdepth: 2
    :caption: Contents:
 
+   source/cli
    source/IonDiff
    source/modules
+   source/settings
+   source/setup
+   source/tests
 
 Authors
 -------
