+++ docs/source/tests.rst
@@ -1,37 +1,0 @@
-tests package
-=============
-
-Submodules
-----------
-
-tests.test\_DIE module
-----------------------
-
-.. automodule:: tests.test_DIE
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_MDR module
-----------------------
-
-.. automodule:: tests.test_MDR
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tests.test\_SPR module
-----------------------
-
-.. automodule:: tests.test_SPR
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: tests
-   :members:
-   :undoc-members:
-   :show-inheritance:
