+++ docs/source/settings.rst
@@ -1,7 +1,0 @@
-settings module
-===============
-
-.. automodule:: settings
-   :members:
-   :undoc-members:
-   :show-inheritance:
