+++ docs/conf.py
@@ -61,7 +61,7 @@ html_theme = "sphinx_book_theme"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-#html_static_path = ["_static"]
+html_static_path = []
 html_favicon = "logo.svg"
 html_logo = "logo.svg"
 #html_theme_options = {
