+++ docs/index.rst
@@ -61,23 +61,6 @@ The IonDiff code can be perfectly used by any scientist performing either classi
 
 There is already a myriad of open-source codes and scripts that can be used for this end, namely, to convert a trajectory file generated by a classical MD code to the VASP format, like, for instance, `LAVA <https://github.com/lanl/LAVA>`_ or `xfroggie <https://www.xfroggie.com/index.cgi/strconv>`_ (to convert from LAMMPS or GROMACS format to VASP format, respectively).
 
-API Documentation
------------------
-
-.. toctree::
-   :hidden:
-   :caption: Modules
-   :maxdepth: 4
-
-   source/modules
-
-.. toctree::
-   :hidden:
-   :caption: Usage
-   :maxdepth: 4
-
-   source/IonDiff
-
 Authors
 -------
 
@@ -91,3 +74,10 @@ Contact, questions and contributing
 -----------------------------------
 
 If you have questions, please don't hesitate to reach out at: cibran.lopez@upc.edu
+
+.. toctree::
+   :hidden:
+   :caption: Usage
+   :maxdepth: 4
+
+   source/IonDiff
