+++ .readthedocs.yaml
@@ -10,11 +10,15 @@ build:
   tools:
     python: "3.11"
 
-# Build from the docs/ directory with Sphinx
-sphinx:
-  configuration: docs/conf.py
-
 # Explicitly set the version of Python and its requirements
 python:
   install:
     - requirements: docs/docs_requirements.txt
+
+# Build from the docs/ directory with Sphinx
+sphinx:
+  configuration: docs/conf.py
+
+commands:
+  - sphinx-apidoc -o docs/source ../IonDiff
+
