+++ .readthedocs.yaml
@@ -18,7 +18,3 @@ python:
 # Build from the docs/ directory with Sphinx
 sphinx:
   configuration: docs/conf.py
-
-commands:
-  - sphinx-apidoc -o docs/source IonDiff
-
