### 2024-07-12 #6616 @IonRepo
Dear Reviewer,

We sincerely appreciate your efforts in reviewing this article, which has
greatly enhanced its quality. The problem that has been raised is based on
the fact that correlations can not be drawn from one single simulation. In
response, and following both Reviewers' advice, we have added an additional
example to the repository. This ensures that all functionalities can now be
run adequately (please see the updated examples folder and README file).
Additionally, we have improved the API documentation (please see the
updated documentation, which now contains the docstrings of the libraries).
We thank the Reviewer for suggesting the Sphinx package, which we have
utilized for this end.

Best regards,
Cibrán.

Missatge de Luis Itza Vazquez-Salazar ***@***.***> del dia
dj., 11 de jul. 2024 a les 14:44:

> Dear Authors,
>
> Thanks for your efforts to improve the code. I am fine with the manuscript
> associated with the code. Nevertheless, I still have two problems with the
> code.
>
>    1. The bug in the first example with analyze_correlations.py in line
>    77 with the number of arguments to unpack. I am using python 3.8.19.
>    Adjuncted a screenshot of the mentioned problem.
>
> Captura.de.pantalla.2024-07-11.a.la.s.14.31.10.png (view on web)
> <https://github.com/openjournals/joss-reviews/assets/59973237/7540ac9a-6b24-4629-965a-d8998eb6c4b4>
>
>    2. Thanks for adding the link to the read-the-docs page with the
>    documentation; it works fine. Nevertheless, to fullfil the requirements of
>    JOSS, you need to add API level documentation. This should not be a problem
>    as all functions in the code are adequately documented, so any automatic
>    API generator like sphinx [
>    https://www.sphinx-doc.org/en/master/man/sphinx-apidoc.html] can do
>    it. This is a minor point, but it will improve the user experience as the
>    arguments of each function can be checked from the documentation instead of
>    reading the source code.
>
> Again, thanks for your effort. Once those two points are sorted, I will
> gladly recommend accepting your work.
>
