### 2024-07-16 #6616 @rkurchin
Thanks everyone! @IonRepo, I'll do an editorial pass over the manuscript and send any comments shortly. In the meantime, the next steps for you are shown in the comment above, so you can work on those checks and send me version number and DOI information whenever you can. Almost there!
