### 2024-07-17 a1786a0
Update paper.md
