### 2024-07-17 #6616 @IonRepo
Dear Editor,

We would like to thank the Editor for their careful remarks; we have
addressed all of them.

In our notation, $n_d$ represents the number of dimensions considered,
while $d$ denotes the specific spatial dimension along which the
ionic-diffusion coefficient is computed. For example, in three dimensions,
$n_d = 3$ while $d$ can be any of the Cartesian directions.

In equation 7, the upper-case I and J are cluster indexes. To improve
readability, we have slightly modified the notation in equations 5-7. For
reference, please see the section on the silhouette coefficient in the
scikit-learn documentation: scikit-learn clustering
<https://scikit-learn.org/stable/modules/clustering.html#silhouette-coefficient>
.
Best regards,
Cibrán.

Missatge de Rachel Kurchin ***@***.***> del dia dt., 16 de
jul. 2024 a les 20:40:

> Editorial remarks:
>
>    - line 11: First sentence of Summary doesn't make sense as written –
>    perhaps "conforming" should be "comprising" instead?
>    - 12: add "techniques" after "analysis"
>    - 13: add hyphen to "high-throughput"
>    - 14: rewrite as "determination of exact ionic migration paths"
>    - 15: "have" -> "has"
>    - 17: add hyphen to "high-performance"
>    - 19: rewrite as "which particles diffuse and when"
>    - 20: "migrating" -> "migration"
>    - 22: remove comma and final clause in sentence (it's clear that
>    you're citing examples)
>    - 30: "simulations" -> "simulation"
>    - 33: rewrite as "can describe in detail the diffusion and vibration
>    of the constituent ions"
>    - 34: replace "painstaking" with "notable" and "handy" with either
>    "user-friendly" or "convenient"
>    - 47: "migrating" -> "migration"
>    - 59: remove extra )
>    - 86, 88, and several equations: subscripts that are not variables
>    (e.g. "self" and "distinct" and also "B" in Boltzmann's constant) should be
>    rendered \mathrm so they're not italicized
>    - 88: "many-atoms" -> "many-atom"
>    - 90: "many-ions" -> "many-ion"
>    - 91: add "and" before "hence"
>    - 93: add "calculation" after "coefficient"
>    - 93: "matricial" -> "matrix"
>    - 94: "like" -> "as"
>    - 99: "diffusses" -> "diffuses"
>    - 101: "tensor" -> "array" and add comma after
>    - 101: remove "very much"
>    - 103: remove "number of"
>    - 105: "four rank tensor" -> "four-dimensional array"
>    - 105: I think $n_d$ should just be $d$ the way you've defined your
>    notation?
>    - 117: This sentence doesn't really make sense, I think perhaps
>    "conforms" should be "constructs" here?
>    - equation 7: what do the upper-case I and J mean?
>
