+++ build/lib/IonDiff/identify_diffusion.py
@@ -1,0 +1,299 @@
+#!/usr/bin/env python
+import numpy             as np
+import matplotlib.pyplot as plt
+import seaborn           as sns
+import sys
+
+from IonDiff import common_library as CL
+
+sns.set_theme()
+
+"""Definition of the class to extract the diffusion information. Only VASP simulations are considered as yet.
+"""
+
+# Defining the class
+class xdatcar:
+    """Python Class for loading information from VASP simulations and identifying diffusions.
+
+    Attributes:
+        kmeans_kwargs   (dict):  Parameters for K-means clustering algorithm.
+        spectral_kwargs (dict):  Parameters for Spectral clustering algorithm.
+        time_step       (float): Time step between consecutive XDATCAR configurations.
+        n_ions          (int):   Number of ions in the simulation.
+
+    Methods:
+        __init__(self, args):
+            Initializes the XDATCAR class.
+    """
+
+    def __init__(self, args):
+        """Initialize the XDATCAR class.
+
+        Args:
+            args: Command line arguments containing MD_path and other parameters.
+
+        Raises:
+            exit: If required files are missing.
+        """
+        
+        # Loading intervals step and number of simulation steps between records
+        self.delta_t, self.n_steps = CL.read_INCAR(args.MD_path)
+        
+        # Time step between consecutive XDATCAR configurations
+        self.time_step = self.n_steps * self.delta_t
+        
+        # Reading the simulation data
+        try:
+            self.read_simulation(args)
+        except:
+            sys.exit('Error: simulation not found')
+
+
+    def read_simulation(self, args):
+        """Reads VASP XDATCAR files.
+        """
+
+        cell, self.n_ions, self.compounds, self.concentration, position = CL.read_XDATCAR(args.MD_path)
+        
+        self.positionC = np.zeros_like(position)
+        self.n_iter    = position.shape[0]
+        
+        # Getting the variation in positions and applying periodic boundary condition
+        dpos = np.diff(position, axis=0)
+        dpos[dpos >  0.5] -= 1.0
+        dpos[dpos < -0.5] += 1.0
+        
+        # Getting the positions and variations in cell units
+        for i in range(self.n_iter-1):
+            self.positionC[i] = np.dot(position[i], cell)
+            dpos[i]           = np.dot(dpos[i],     cell)
+
+        self.positionC[-1] = np.dot(position[-1], cell)
+        self.dpos          = dpos
+
+        expanded_dimensions        = np.expand_dims(self.positionC[0], 0)
+        self.cartesian_coordinates = np.concatenate([expanded_dimensions, self.dpos], axis=0)
+        self.cartesian_coordinates = np.cumsum(self.cartesian_coordinates, axis=0)
+        
+        # Defining the attribute of window=1 variation in position and velocity
+        self.velocity = self.dpos / self.time_step
+
+    
+    def get_diffusion(self, args):
+        """Obtains diffusion information from the simulation data.
+
+        Args:
+            args: Command line arguments.
+
+        Returns:
+            list: List of diffusion events, each represented as [particle index, start step, end step].
+        """
+        
+        hoppings = []
+        if args.make_plot:
+            fig = plt.figure(figsize=(5, 5))
+            ax = plt.axes(projection='3d')
+            ax.set_xlabel('x')
+            ax.set_ylabel('y')
+            ax.set_zlabel('z')
+        
+        for particle in range(self.n_ions):
+            coordinates = self.cartesian_coordinates[:, particle]
+            
+            # Recommended number of clusters
+            n_clusters = CL.calculate_silhouette(coordinates,
+                                                 args.classifier,
+                                                 args.n_attempts,
+                                                 args.silhouette_thd)
+            
+            # Information from the clustering
+            centers, classification, vibration, cluster_change = CL.calculate_clusters(coordinates,
[…]
