+++ setup.py
@@ -6,7 +6,7 @@ with open('LICENSE') as f:
 setup(
   name = 'IonDiff',
   packages = ['IonDiff'],
-  version = '1.5',
+  version = '1.6',
   license=license,
   description = 'Unsupervised identification and analysis of ion-hopping events in solid state electrolytes.',
   author = 'Cibrán López Álvarez',
