### 2024-07-18 #6616 @IonRepo
Dear Editor,

This would be the release 1.6 with DOI: 10.5281/zenodo.12771455.

Best regards,
Cibrán.

Missatge de Rachel Kurchin ***@***.***> del dia dc., 17 de
jul. 2024 a les 18:47:

> @IonRepo <https://github.com/IonRepo>, awesome, manuscript looks great.
> So all we need is the version number of the release with all the changes
> from the review and the DOI of the deposit, then we should be ready to go!
>
