### 2024-07-18 #6616 @IonRepo
Dear Editor,

Do you mean the title and author list that appears in the README.md file of
the repository?

As well, I am not quite sure how to modify the license in the draft, as it
seems to be predetermined by the JOSS template (I did not explicit the CC
license in the draft).

Best regards,
Cibrán.

Missatge de Rachel Kurchin ***@***.***> del dia dj., 18 de
jul. 2024 a les 16:00:

> @IonRepo <https://github.com/IonRepo>, can you please make sure that the
> title and author list of the archive match the paper?
>
> Also, the repo seems to have an MIT license, while the archive is CC, so
> please make those consistent as well.
>
