### 2024-07-18 #6616 @rkurchin
I'm referring to the title of the Zenodo archive.

Similarly, for the license, I'm talking about the license listed on Zenodo vs. the one in your repository; they do not currently match each other.
