### 2024-07-18 #6616 @IonRepo
Hi again,

Thank you, all the information has been updated in Zenodo accordingly.

Best regards,
Cibrán.

Missatge de Rachel Kurchin ***@***.***> del dia dj., 18 de
jul. 2024 a les 17:09:

> I'm referring to the title of the Zenodo archive.
>
> Similarly, for the license, I'm talking about the license listed on Zenodo
> vs. the one in your repository; they do not currently match each other.
>
