### 2024-07-19 #6616 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
