### 2024-07-21 #6616 @Kevin-Mattheus-Moerman
@IonRepo @Claudiator7 as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, however the below are some points that require your attention:

On the paper:
- [ ]  Both British English (e.g. `analysing`) and American English (e.g. the use of the `z` in `optimized`, `minimize`, `maximized`, and `vectorization`) are used. Please use one consistently and edit the paper accordingly.
