+++ README.md
@@ -6,7 +6,7 @@ Despite playing a central role in the design of high performance solid-state ele
 
 However, the identification of hopping events in computer simulations typically relies on active supervision and definition of arbitrary material-dependent geometrical parameters, thus frustrating high throughput screenings of diffusing paths and mechanisms across simulation databases and the assessment of many-diffusing-ion correlations.   
 
-Here, we introduce a novel approach for analysing ion hopping events in molecular dynamics (MD) simulations in a facile and totally unsupervised manner, what would allow the extraction of completely new descriptors related to these diffusions. Our approach relies on the k-means clustering algorithm and allows to identify with precision which and when particles diffuse in a simulation and the exact migrating paths that they follow as well.
+Here, we introduce a novel approach for analyzing ion hopping events in molecular dynamics (MD) simulations in a facile and totally unsupervised manner, what would allow the extraction of completely new descriptors related to these diffusions. Our approach relies on the k-means clustering algorithm and allows to identify with precision which and when particles diffuse in a simulation and the exact migrating paths that they follow as well.
 
 Documentation showing functionality and usage of the code are provided one the [`docs`](https://iondiff.readthedocs.io/en/latest/) site. Please be aware that the code is under active development, bug reports are welcomed in the GitHub issues!
 
